+++ src/managers/roommanager/CMakeLists.txt
@@ -1,7 +1,7 @@
 cmake_minimum_required(VERSION 3.0.0)
 
 project(
-    suitemanager
+    roommanager
     VERSION 0.1.0
     DESCRIPTION "Manager that handles Rooms. Room has its background texture and ambient animations. Also a collection of game object
     the player can interact with. Suite manager mediates this interaction"
