+++ src/managers/roommanager/src/Ambient.cpp
