+++ src/managers/roommanager/src/Room.cpp
