+++ src/managers/roommanager/src/RoomManager.cpp
